Update domain models

+++ src/main/java/edu/mum/coffee/service/OrderService.java
@@ -37,7 +37,7 @@ public class OrderService {
         return orderRepository.findOrderByOrderDateBetween(minDate, maxDate);
     }
 
-    public Order findById(int id) {
+    public Order findById(Long id) {
         return orderRepository.findOne(id);
     }
